new contract form validation can't delete purchase if he has a contract number switch in contract form refresh grid after delete dialog confirm delete buttons switch

+++ src/main/java/io/mkolodziejczyk92/data/controllers/PurchasesViewController.java
@@ -42,16 +42,24 @@ public class PurchasesViewController {
         }
     }
 
-    public void deletePurchase(Purchase purchase) {
+    public boolean deletePurchase(Purchase purchase) {
         try {
-            purchaseService.delete(purchase.getId());
+            if(purchase.getContractNumber() == null || purchase.getContractNumber().isBlank()){
+                purchaseService.delete(purchase.getId());
+            }else {
+                Notification.show("Purchase "
+                        + purchase.getId()
+                        + " cannot be deleted because it has connections in the database.");
+                return false;
+            }
         } catch (DataIntegrityViolationException e) {
             Notification.show("Purchase "
                     + purchase.getId()
                     + " cannot be deleted because it has connections in the database.");
-            return;
+            return false;
         }
         Notification.show("Purchase " + purchase.getId() + " deleted.");
+        return true;
     }
 
     public List<Purchase> clientPurchases(Long clientId) {
